src/console-sk/requirements.txt
azure-cosmos>=4.0.0
azure-identity
python-dotenv
opentelemetry-api
opentelemetry-sdk
azure-monitor-events-extension

semantic-kernel[azure]==1.32.2
azure-ai-projects==1.0.0b11
openai==1.84.0 
azure-ai-inference==1.0.0b9 
azure-search-documents 
azure-ai-evaluation

# Testing tools (optional)
pytest>=8.2,<9
pytest-asyncio==0.24.0
pytest-cov==5.0.0
